Smells, Sounds, and Touch
Made all of the rooms in Ms. Light’s house. There was the problem of
adding in the different senses to each room. After a difficult time, I
was able to add in smells, sounds and touch as an alternative to
looking and examining all the time. Also, finished prologue. Got rid of
the problem of entering through doors by getting rid of doors
altogether.