Descriptions + Lots of Things added for puzzle
Started the set up of all of the keys on the piano for the puzzle to
work. I made the description of the east room (basement) along with the
touches, smells, sounds, and this time TASTES added to the things in
the room. Edited prologue, and fixed spelling. Having trouble with
setting up the puzzle so that when all of the keys are pressed the
lights come back on. Need to figure out how to trigger that event.